% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chunked-Read.R
\name{chunk_reader}
\alias{chunk_reader}
\title{Gradually read a file from disk}
\usage{
chunk_reader(file_path)
}
\arguments{
\item{file_path}{path to large file}
}
\value{
a list-object containing the function \code{read} to
read lines from the given file, and \code{close} to close the
connection to the file stream.
}
\description{
This function reads a text file in batches (or chunks).
It can be used for very large files that cannot fit in RAM.
}
