% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chunked-Read.R
\name{create_table}
\alias{create_table}
\title{create table}
\usage{
create_table(con, tablename, names, types)
}
\arguments{
\item{con}{DBI connection to database}

\item{tablename}{name of the table of interest}

\item{names}{name of the columns of the table}

\item{types}{sql types of names (in the same order)}
}
\value{
result of the creation query
}
\description{
create a table on a remote database using
defined names and types
}
