% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_structure_lexer_parser.R
\name{load_structure}
\alias{load_structure}
\title{Lex and parse table structure}
\usage{
load_structure(structure, from.file = FALSE)
}
\arguments{
\item{structure}{the text describing the table}

\item{from.file}{if TRUE, input is interpretes as a path}
}
\value{
nested list describing the table
}
\description{
Creates a nested lists representing the structure the complex table
that we want to simplify and load into a database.
}
\examples{
load_structure('./inst/extdata/test.treatments')

}
