% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAF-Rreader.R
\name{maf_db_loader}
\alias{maf_db_loader}
\title{Create a full db from a maf file}
\usage{
maf_db_loader(file_path, table.name, names, types)
}
\arguments{
\item{file_path}{path to MAF file}

\item{table.name}{name of the main db table}

\item{names}{names of the columns (even non exhaustive and in any order)}

\item{types}{types of the columns}
}
\value{
a list object (see code)
}
\description{
This procedures prepares the structures to load a MAF file into
a structured database. The default structure of the databased is
based on GDC standard.
}
