% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maf_db_reader}
\alias{maf_db_reader}
\title{Prepare queries to store a maf file in a database}
\usage{
maf_db_reader(table_name, text, header, rules, starting_point)
}
\arguments{
\item{table_name}{name of the db table}

\item{text}{group of maf lines (original text)}

\item{header}{names of the columns}

\item{rules}{list of actions to manage fields (quoting)}

\item{starting_point}{starting index for "db_index" column (primary key)

Function exported for R}
}
\value{
the necessary insertion query to load MAF data into a database
}
\description{
-- tested with PostgreSQL --
}
