% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAFdb_object.R
\docType{class}
\name{MAFdb-class}
\alias{MAFdb-class}
\title{This clasd is used to create an intuitive interface to a database containing
MAF file data (possibly in GDC format)}
\description{
(In the current form, using MAFdb constructor, can operate on any database
to simplify access to tabls and columns)
}
\section{Slots}{

\describe{
\item{\code{con}}{the DBI connection to the database}
}}

