% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAFdb_object.R
\name{MAFdb.load}
\alias{MAFdb.load}
\title{Create a MAFdb from file}
\usage{
MAFdb.load(
  con,
  path,
  names = NULL,
  types = NULL,
  limit = NULL,
  max_chunk = 10000,
  reset = FALSE
)
}
\arguments{
\item{con}{connection to the database}

\item{path}{path to the MAF file}

\item{names}{list of column names (to specify type)}

\item{types}{list of types associated to column names}

\item{limit}{maximum number of lines to be read}

\item{max_chunk}{maximum number of lines to be read in a single pass}

\item{reset}{Drop the current dataset and make new tables?}
}
\value{
a MAFdb object
}
\description{
Creates a MAFdb directly from a MAF file. This procedure
loads the file into the database gradually, in chuncks, in
order to reduce RAM usage.
}
