% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAFdb_object.R
\name{print,MAFdb-method}
\alias{print,MAFdb-method}
\title{Print tables and columns of this db}
\usage{
\S4method{print}{MAFdb}(x)
}
\arguments{
\item{x}{a MAFdb object}
}
\description{
Print tables and columns of this db
}
