% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAFdb_object.R
\name{pull.indexes.from.table}
\alias{pull.indexes.from.table}
\title{Get all information on these indexes}
\usage{
\method{pull}{indexes.from.table}(data.flow, table)
}
\arguments{
\item{data.flow}{dbplyr data flow from the original database}

\item{maf.db}{a db table (with db_index column)}
}
\value{
a data frame with the resulting information
}
\description{
Creates a tibble representing the information on the currently selected indexes
from the selected table.
This procedure loads results in RAM, use with care.
}
