% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAFdb.R
\docType{package}
\name{rMAFdb}
\alias{rMAFdb}
\title{rMAFdb Package}
\description{
This package creates an easy to use interface to operate with huge MAF files.
MAF filed (possibly following GDC standard) can be parsed and loaded into a database
directly, data manipulation can be then done with dbplyr so that the related
code can be easly adapted from regular "on RAM" operations.
The use of databases allows the user to create indexes to esily speed up queries
even for smaller data sets.
}
\author{
Nicolò Rossi \email{olocin.issor@gmail.com}
}
