% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAFdb_object.R
\name{show,MAFdb-method}
\alias{show,MAFdb-method}
\title{Print tables and columns of this db}
\usage{
\S4method{show}{MAFdb}(object)
}
\arguments{
\item{object}{this MAFdb}
}
\description{
Print tables and columns of this db
}
