% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMAFdb_object.R
\name{[,MAFdb,ANY,missing-method}
\alias{[,MAFdb,ANY,missing-method}
\title{Access a specific table of this database}
\usage{
\S4method{[}{MAFdb,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a MAFdb object}

\item{i}{name of the table to be accessed}

\item{j}{unused}

\item{...}{unused}

\item{drop}{unused}
}
\description{
Access a specific table of this database
}
