% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_MAFdb}
\alias{test_MAFdb}
\title{Test}
\usage{
test_MAFdb(table_name, text, header, rules, starting_point)
}
\arguments{
\item{table_name}{name of the db table}

\item{text}{group of maf lines (original text)}

\item{header}{names of the columns}

\item{rules}{list of actions to manage fields (quoting)}

\item{starting_point}{starting index for "db_index" column (primary key)

Function exported for R}
}
\value{
the necessary insertion query to load MAF data into a database
}
\description{
Simple testing procedure used for a small table and to show functionalities
(see .Rmd file)
}
